fix flaky test relying on application env
